+++ deltafi-core/src/main/java/org/deltafi/core/configuration/StorageCacheConfig.java
@@ -15,18 +15,19 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.deltafi.core;
+package org.deltafi.core.configuration;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.retry.annotation.EnableRetry;
-import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.cache.CacheManager;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cache.concurrent.ConcurrentMapCacheManager;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 
-@SpringBootApplication
-@EnableAsync
-@EnableRetry
-public class DeltaFiCoreApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(DeltaFiCoreApplication.class, args);
-	}
+@Configuration
+@EnableCaching
+public class StorageCacheConfig {
+    @Bean
+    public CacheManager cacheManager() {
+        return new ConcurrentMapCacheManager("diskspaceservice-storage");
+    }
 }